+++ app/components/asset/AddModal.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import * as z from 'zod'
 import type { FormSubmitEvent } from '@nuxt/ui'
 
 import { useCategory } from '~/composables/useCategory'
@@ -10,79 +9,19 @@ import { useLocation } from '~/composables/useLocation'
 import { useAssetLocation } from '~/composables/useAssetLocation'
 import { useBranch } from '~/composables/useBranch'
 import { formatCalendarDate, calendarDateToISOString } from '~/utils/date'
+import {
+  createAssetSchema as schema,
+  categorySchema,
+  subCategorySchema,
+  assetLocationSchema as locationSchema,
+  type CreateAssetSchema as Schema,
+  type CategorySchema,
+} from '~/schemas/assetSchema'
 
 const props = defineProps<{
   initialCode?: string
 }>()
 
-const schema = z.object({
-  codes: z.array(z.string().min(1, 'Asset code is required')).min(1, 'At least one code is required'),
-  name: z.string().min(1, 'Asset name is required'),
-  description: z.string().optional(),
-  brand: z.string().optional(),
-  model: z.string().optional(),
-  user: z.string().min(1, 'User is required'),
-  price: z.union([z.string(), z.number()]).refine(val => val !== undefined && val !== '', {
-    message: 'Price is required'
-  }),
-  purchaseDate: z.string().min(1, 'Purchase date is required'),
-  categoryId: z.string().min(1, 'Category is required'),
-  subCategoryId: z.string().min(1, 'Sub category is required'),
-  locationId: z.string().optional(),
-  isLendable: z.boolean().default(false),
-  image: z.any().refine(file => file !== null && file !== undefined, {
-    message: 'Asset image is required'
-  }),
-  properties: z.array(
-    z.object({
-      id: z.string(),
-      value: z.union([z.string(), z.number()]).optional()
-    })
-  ).optional(),
-  labels: z.array(
-    z.object({
-      key: z.string().min(1, 'Label key is required'),
-      value: z.string().min(1, 'Label value is required')
-    })
-  ).optional()
-}).superRefine((data, ctx) => {
-  if (data.labels) {
-    const keys = new Map<string, number[]>()
-    data.labels.forEach((item, index) => {
-      const k = item.key.toLowerCase().trim()
-      if (k) {
-        if (!keys.has(k)) keys.set(k, [])
-        keys.get(k)!.push(index)
-      }
-    })
-  }
-})
-type Schema = z.output<typeof schema>
-
-const categorySchema = z.object({
-  name: z.string().min(1, 'Name is required'),
-  hasMaintenance: z.boolean().default(false),
-  hasHolder: z.boolean().default(false),
-  hasLocation: z.boolean().default(false)
-})
-
-const subCategorySchema = z.object({
-  name: z.string().min(1, 'Name is required'),
-  categoryId: z.string().min(1, 'Category is required'),
-  properties: z.array(
-    z.object({
-      name: z.string().min(1, 'Property name is required'),
-      dataType: z.enum(['string', 'number'], { message: 'Type is required' })
-    })
-  )
-})
-
-const locationSchema = z.object({
-  name: z.string().min(1, 'Name is required'),
-  branchId: z.string().min(1, 'Branch is required')
-})
-
-type CategorySchema = z.infer<typeof categorySchema>
 
 const emit = defineEmits<{ (e: 'created'): void }>()
